requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.8.1
async-timeout==4.0.2
attrs==23.1.0
autopep8==2.0.2
beautifulsoup4==4.12.2
charset-normalizer==3.1.0
cryptography==43.0.1
Django==5.1.1
django-cors-headers==4.3.1
django-extensions==3.2.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.1.2
frozenlist==1.3.3
google-generativeai==0.8.5
httpx==0.27.0
idna==3.4
iniconfig==2.0.0
kafka-python===2.0.2
multidict==6.0.4
numpy
packaging==23.1
pluggy==1.2.0
psycopg2-binary==2.9.9
pycodestyle
pytest==7.4.0
python-dateutil==2.9.0.post0
python-decouple==3.8
requests==2.31.0
sqlparse==0.5.1
tomli==2.0.1
whitenoise==6.8.2
yarl==1.9.2
psutil==7.0.0
pytz==2025.2
cryptography==43.0.1